# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VIGORA - Plataforma de Gestão Jurídica
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clientes:
    get:
      operationId: listClientes
      tags: [clientes]
      summary: List clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
    post:
      operationId: createCliente
      tags: [clientes]
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClienteRequest"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
  /clientes/{id}:
    get:
      operationId: getCliente
      tags: [clientes]
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClienteWithRelations"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCliente
      tags: [clientes]
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClienteRequest"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
    delete:
      operationId: deleteCliente
      tags: [clientes]
      summary: Delete client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Client deleted
...
  /alertas:
    get:
      operationId: listAlertas
      tags: [alertas]
      summary: List alerts
      parameters:
        - name: lido
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alerta"
  /alertas/{id}/lido:
    put:
      operationId: marcarAlertaLido
      tags: [alertas]
      summary: Mark alert as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alerta"
...
  /contatos-whatsapp:
    get:
      operationId: listContatosWhatsapp
      tags: [contatos-whatsapp]
      summary: List WhatsApp contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContatoWhatsapp"
    post:
      operationId: createContatoWhatsapp
      tags: [contatos-whatsapp]
      summary: Create WhatsApp contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContatoWhatsappRequest"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContatoWhatsapp"
        "409":
          description: Duplicate number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contatos-whatsapp/{id}:
    get:
      operationId: getContatoWhatsapp
      tags: [contatos-whatsapp]
      summary: Get contact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContatoWhatsapp"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateContatoWhatsapp
      tags: [contatos-whatsapp]
      summary: Update WhatsApp contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContatoWhatsappRequest"
      responses:
        "200":
          description: Contact updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContatoWhatsapp"
    delete:
      operationId: deleteContatoWhatsapp
      tags: [contatos-whatsapp]
      summary: Delete WhatsApp contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Contact deleted
...
  /notificacoes-processos/{processoId}/contatos:
    get:
      operationId: listContatosDoProcesso
      tags: [notificacoes-processos]
      summary: List WhatsApp contacts linked to a process
      parameters:
        - name: processoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of linked contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContatoWhatsapp"
    post:
      operationId: salvarVinculosProcesso
      tags: [notificacoes-processos]
      summary: Set (sync) WhatsApp contacts linked to a process
      parameters:
        - name: processoId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalvarVinculosRequest"
      responses:
        "200":
          description: Links saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VinculoSaveResult"
  /notificacoes-processos/{processoId}/contatos/{contatoId}:
    delete:
      operationId: removerContatoDoProcesso
      tags: [notificacoes-processos]
      summary: Remove a WhatsApp contact link from a process
      parameters:
        - name: processoId
          in: path
          required: true
          schema:
            type: integer
        - name: contatoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Link removed
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        senha:
          type: string
      required:
        - email
        - senha
    RegisterRequest:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
          format: email
        telefone:
          type: string
        senha:
          type: string
      required:
        - nome
        - email
        - senha
    User:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
          nullable: true
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - nome
        - email
        - criadoEm
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    Cliente:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        cpfCnpj:
          type: string
          nullable: true
        telefone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        observacoes:
          type: string
          nullable: true
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - nome
        - criadoEm
    ClienteWithRelations:
      allOf:
        - $ref: "#/components/schemas/Cliente"
        - type: object
          properties:
            processos:
              type: array
              items:
                $ref: "#/components/schemas/ProcessoSummary"
            prazos:
              type: array
              items:
                $ref: "#/components/schemas/PrazoSummary"
          required:
            - processos
            - prazos
    CreateClienteRequest:
      type: object
      properties:
        nome:
          type: string
        cpfCnpj:
          type: string
          nullable: true
        telefone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        observacoes:
          type: string
          nullable: true
      required:
        - nome
...
    ProcessoSummary:
      type: object
      properties:
        id:
          type: integer
        numeroProcesso:
          type: string
        area:
          type: string
          nullable: true
        status:
          type: string
      required:
        - id
        - numeroProcesso
        - status
...
    PrazoSummary:
      type: object
      properties:
        id:
          type: integer
        tipo:
          type: string
        dataLimite:
          type: string
          format: date
        status:
          type: string
        prioridade:
          type: string
      required:
        - id
        - tipo
        - dataLimite
        - status
        - prioridade
...
    Alerta:
      type: object
      properties:
        id:
          type: integer
        tipo:
          type: string
          enum: [prazo_vencido, prazo_urgente, prazo_criado, tarefa_atrasada, tarefa_concluida, processo_atencao]
        mensagem:
          type: string
        referenciaId:
          type: integer
          nullable: true
        referenciaTipo:
          type: string
          nullable: true
        lido:
          type: boolean
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - tipo
        - mensagem
        - lido
        - criadoEm
...
    ContatoWhatsapp:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        numeroWhatsapp:
          type: string
        email:
          type: string
          nullable: true
        status:
          type: string
          enum: [ativo, inativo]
        observacoes:
          type: string
          nullable: true
        usuarioId:
          type: integer
          nullable: true
        criadoEm:
          type: string
          format: date-time
        atualizadoEm:
          type: string
          format: date-time
      required:
        - id
        - nome
        - numeroWhatsapp
        - status
        - criadoEm
        - atualizadoEm
    CreateContatoWhatsappRequest:
      type: object
      properties:
        nome:
          type: string
        numeroWhatsapp:
          type: string
        email:
          type: string
          nullable: true
        status:
          type: string
          enum: [ativo, inativo]
        observacoes:
          type: string
          nullable: true
      required:
        - nome
        - numeroWhatsapp
...
    SalvarVinculosRequest:
      type: object
      properties:
        contatoIds:
          type: array
          items:
            type: integer
      required:
        - contatoIds
    VinculoSaveResult:
      type: object
      properties:
        success:
          type: boolean
        added:
          type: integer
        removed:
          type: integer
      required:
        - success
        - added
        - removed