# lib/api-zod/src/generated/types/alerta.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIGORA - Plataforma de Gestão Jurídica
 * OpenAPI spec version: 0.1.0
 */
import type { AlertaTipo } from "./alertaTipo";

export interface Alerta {
  id: number;
  tipo: AlertaTipo;
  mensagem: string;
  referenciaId?: number | null;
  referenciaTipo?: string | null;
  lido: boolean;
  criadoEm: Date;
}