# lib/api-zod/src/generated/types/contatoWhatsapp.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIGORA - Plataforma de Gestão Jurídica
 * OpenAPI spec version: 0.1.0
 */
import type { ContatoWhatsappStatus } from "./contatoWhatsappStatus";

export interface ContatoWhatsapp {
  id: number;
  nome: string;
  numeroWhatsapp: string;
  email?: string | null;
  status: ContatoWhatsappStatus;
  observacoes?: string | null;
  usuarioId?: number | null;
  criadoEm: Date;
  atualizadoEm: Date;
}