# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIGORA - Plataforma de Gestão Jurídica
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  nome: string;
  email: string;
  telefone?: string | null;
  criadoEm: Date;
}